+++ src/App.tsx
@@ -136,13 +136,6 @@ export interface ConfirmModalState {
   title?: string;
 }
 
-interface Metadata {
-  adjustments: Adjustments;
-  rating: number;
-  tags: Array<string> | null;
-  version: number;
-}
-
 export interface MultiSelectOptions {
   onSimpleClick(p: any): void;
   updateLibraryActivePath: boolean;
@@ -202,26 +195,16 @@ export interface SearchCriteria {
 function App() {
   const {
     selectedImage,
-    setSelectedImage,
     appSettings,
     rootPath,
-    setRootPath,
     activeView,
     setActiveView,
     isWindowFullScreen,
     isLayoutReady,
-    setIsLayoutReady,
     currentFolderPath,
-    setCurrentFolderPath,
-    expandedFolders,
-    setExpandedFolders,
     folderTree,
-    setFolderTree,
-    setPinnedFolderTrees,
     imageList,
-    setImageList,
     imageRatings,
-    setImageRatings,
     sortCriteria,
     filterCriteria,
     supportedTypes,
@@ -230,37 +213,18 @@ function App() {
     libraryActivePath,
     setLibraryActivePath,
     libraryActiveAdjustments,
-    setLibraryActiveAdjustments,
-    setFinalPreviewUrl,
-    setUncroppedAdjustedPreviewUrl,
     setShowOriginal,
     adjustments,
-    setAdjustments: setLiveAdjustments,
     isTreeLoading,
-    setIsTreeLoading,
     isViewLoading,
-    setIsViewLoading,
-    initialFileToOpen,
-    setInitialFileToOpen,
-    setError,
-    setHistogram,
-    waveform,
-    setWaveform,
-    isWaveformVisible,
     setIsWaveformVisible,
     uiVisibility,
     setUiVisibility,
-    isSliderDragging,
     setIsSliderDragging,
     isFullScreen,
-    setIsFullScreen,
-    isHighResNeeded,
-    setIsHighResNeeded,
     isAnimatingTheme,
     theme,
-    dragIdleTimer,
     activeRightPanel,
-    setActiveRightPanel,
     slideDirection,
     activeMaskContainerId,
     setActiveMaskContainerId,
@@ -271,19 +235,11 @@ function App() {
     activeAiSubMaskId,
     setActiveAiSubMaskId,
     zoom,
-    setZoom,
     displaySize,
-    previewSize,
-    setPreviewSize,
     baseRenderSize,
     originalSize,
-    setOriginalSize,
-    setIsLoadingFullRes,
-    fullResCacheKeyRef,
-    initialFitScale,
     setInitialFitScale,
     renderedRightPanel,
-    setRenderedRightPanel,
     isLibraryExportPanelVisible,
     setIsLibraryExportPanelVisible,
     libraryViewMode,
@@ -293,87 +249,50 @@ function App() {
     setRightPanelWidth,
     bottomPanelHeight,
     setBottomPanelHeight,
-    setActiveTreeSection,
     isResizing,
     thumbnailSize,
     setThumbnailSize,
     thumbnailAspectRatio,
     setThumbnailAspectRatio,
     copiedAdjustments,
-    setCopiedAdjustments,
     isStraightenActive,
     setIsStraightenActive,
     copiedFilePaths,
     setCopiedFilePaths,
     aiModelDownloadStatus,
-    setAiModelDownloadStatus,
     isCopied,
-    setIsCopied,
     isPasted,
-    setIsPasted,
     isIndexing,
-    setIsIndexing,
     indexingProgress,
-    setIndexingProgress,
     searchCriteria,
-    setSearchCriteria,
-    isCreateFolderModalOpen,
     setIsCreateFolderModalOpen,
-    isRenameFolderModalOpen,
     setIsRenameFolderModalOpen,
-    isRenameFileModalOpen,
     setIsRenameFileModalOpen,
-    renameTargetPaths,
-    setRenameTargetPaths,
-    isImportModalOpen,
     setIsImportModalOpen,
-    isCopyPasteSettingsModalOpen,
     setIsCopyPasteSettingsModalOpen,
-    importTargetFolder,
-    setImportTargetFolder,
     importSourcePaths,
-    setImportSourcePaths,
     folderActionTarget,
-    setFolderActionTarget,
     confirmModalState,
-    setConfirmModalState,
-    panoramaModalState,
     setPanoramaModalState,
-    hdrModalState,
     setHdrModalState,
     negativeModalState,
     setNegativeModalState,
-    denoiseModalState,
     setDenoiseModalState,
-    collageModalState,
     setCullingModalState,
-    cullingModalState,
     setCollageModalState,
     customEscapeHandler,
     libraryScrollTop,
-    setLibraryScrollTop,
     thumbnails,
     setThumbnails,
-    isInitialMount,
-    isProgrammaticZoom,
-    currentFolderPathRef,
-    preloadedDataRef,
-    previewJobIdRef,
-    latestRenderedJobIdRef,
     history,
-    exportState,
-    setExportState,
     importState,
-    setImportState,
     isLightTheme,
-    visualAdjustmentsKey,
-    pinnedFolders,
     isAnyModalOpen,
   } = useAppState();
 
   const { sortedImageList } = useSortedImageList();
 
-  const { canUndo, canRedo, resetHistory: resetAdjustmentsHistory, goToIndex: goToAdjustmentsHistoryIndex } = history;
+  const { canUndo, canRedo, goToIndex: goToAdjustmentsHistoryIndex } = history;
 
   const {
     handleBackToLibrary,
@@ -381,10 +300,7 @@ function App() {
     executeDelete,
     handleDeleteSelected,
     refreshImageList,
-    applyAdjustments,
-    debouncedSave,
     handleDisplaySizeChange,
-    debouncedSetHistory,
     setAdjustments,
     handleStraighten,
     toggleWbPicker,
@@ -405,9 +321,7 @@ function App() {
     handleRightPanelSelect,
     handleSettingsChange,
     handleToggleWaveform,
-    refreshAllFolderTrees,
     handleActiveTreeSectionChange,
-    handleTogglePinFolder,
     handleSelectSubfolder,
     handleLibraryRefresh,
     handleToggleFolder,
@@ -417,12 +331,9 @@ function App() {
     handleAutoAdjustments,
     handleRate,
     handleSetColorLabel,
-    getCommonTags,
     handleTagsChanged,
     closeConfirmModal,
     handlePasteFiles,
-    requestFullResolution,
-    handleFullResolutionLogic,
     handleZoomChange,
     handleUserTransform,
     handleSavePanorama,
@@ -433,20 +344,22 @@ function App() {
     handleOpenFolder,
     handleContinueSession,
     handleGoHome,
-    handleMultiSelectClick,
     handleLibraryImageSingleClick,
     handleImageClick,
     handleClearSelection,
-    handleRenameFiles,
     handleSaveRename,
     handleStartImport,
     handleResetAdjustments,
-    handleImportClick,
+    handleEditorContextMenu,
+    handleThumbnailContextMenu,
+    handleCreateFolder,
+    handleRenameFolder,
+    handleFolderTreeContextMenu,
+    handleMainLibraryContextMenu,
   } = useHandlers();
 
   useGlobalEffects();
 
-  const { showContextMenu } = useContextMenu();
   const { loading: isThumbnailsLoading } = useThumbnails(imageList, setThumbnails);
 
   useKeyboardShortcuts({
@@ -496,779 +409,6 @@ function App() {
     baseRenderSize,
     originalSize,
   });
-
-  const handleEditorContextMenu = (event: any) => {
-    event.preventDefault();
-    event.stopPropagation();
-    if (!selectedImage) return;
-
-    const handleCreateVirtualCopy = async (sourcePath: string) => {
-      try {
-        await invoke(Invokes.CreateVirtualCopy, { sourceVirtualPath: sourcePath });
-        await refreshImageList();
-      } catch (err) {
-        console.error('Failed to create virtual copy:', err);
-        setError(`Failed to create virtual copy: ${err}`);
-      }
-    };
-
-    const commonTags = getCommonTags([selectedImage.path]);
-
-    const options: Array<Option> = [
-      {
-        label: 'Export Image',
-        icon: Save,
-        onClick: () => {
-          setRenderedRightPanel(Panel.Export);
-          setActiveRightPanel(Panel.Export);
-        },
-      },
-      { type: OPTION_SEPARATOR },
-      { label: 'Undo', icon: Undo, onClick: undo, disabled: !canUndo },
-      { label: 'Redo', icon: Redo, onClick: redo, disabled: !canRedo },
-      { type: OPTION_SEPARATOR },
-      { label: 'Copy Adjustments', icon: Copy, onClick: handleCopyAdjustments },
-      {
-        label: 'Paste Adjustments',
-        icon: ClipboardPaste,
-        onClick: handlePasteAdjustments,
-        disabled: copiedAdjustments === null,
-      },
-      {
-        label: 'Productivity',
-        icon: Gauge,
-        submenu: [
-          {
-            label: 'Auto Adjust Image',
-            icon: Aperture,
-            onClick: handleAutoAdjustments,
-          },
-          {
-            icon: CopyPlus,
-            label: 'Create Virtual Copy',
-            onClick: () => handleCreateVirtualCopy(selectedImage.path),
-          },
-          {
-            label: 'Denoise',
-            icon: Grip,
-            onClick: () => {
-              setDenoiseModalState({
-                isOpen: true,
-                isProcessing: false,
-                previewBase64: null,
-                error: null,
-                targetPath: selectedImage.path,
-                progressMessage: null,
-              });
-            },
-          },
-          {
-            label: 'Convert Negative',
-            icon: Film,
-            onClick: () => {
-              if (selectedImage) {
-                setNegativeModalState({
-                  isOpen: true,
-                  targetPath: selectedImage.path,
-                });
-              }
-            },
-          },
-          {
-            disabled: true,
-            icon: SquaresUnite,
-            label: 'Stitch Panorama',
-          },
-          {
-            disabled: true,
-            icon: Images,
-            label: 'Merge to HDR',
-          },
-          {
-            icon: LayoutTemplate,
-            label: 'Frame Image',
-            onClick: () => {
-              setCollageModalState({
-                isOpen: true,
-                sourceImages: [selectedImage],
-              });
-            },
-          },
-          {
-            label: 'Cull Image',
-            icon: Users,
-            disabled: true,
-          },
-        ],
-      },
-      { type: OPTION_SEPARATOR },
-      {
-        label: 'Rating',
-        icon: Star,
-        submenu: [0, 1, 2, 3, 4, 5].map((rating: number) => ({
-          label: rating === 0 ? 'No Rating' : `${rating} Star${rating !== 1 ? 's' : ''}`,
-          onClick: () => handleRate(rating),
-        })),
-      },
-      {
-        label: 'Color Label',
-        icon: Palette,
-        submenu: [
-          { label: 'No Label', onClick: () => handleSetColorLabel(null) },
-          ...COLOR_LABELS.map((label: Color) => ({
-            label: label.name.charAt(0).toUpperCase() + label.name.slice(1),
-            color: label.color,
-            onClick: () => handleSetColorLabel(label.name),
-          })),
-        ],
-      },
-      {
-        label: 'Tagging',
-        icon: Tag,
-        submenu: [
-          {
-            customComponent: TaggingSubMenu,
-            customProps: {
-              paths: [selectedImage.path],
-              initialTags: commonTags,
-              onTagsChanged: handleTagsChanged,
-              appSettings,
-            },
-          },
-        ],
-      },
-      { type: OPTION_SEPARATOR },
-      {
-        label: 'Reset Adjustments',
-        icon: RotateCcw,
-        onClick: () => {
-          debouncedSetHistory.cancel();
-          const currentRating = adjustments.rating;
-
-          const originalAspectRatio =
-            selectedImage.width && selectedImage.height ? selectedImage.width / selectedImage.height : null;
-
-          resetAdjustmentsHistory({
-            ...INITIAL_ADJUSTMENTS,
-            aspectRatio: originalAspectRatio,
-            rating: currentRating,
-            aiPatches: [],
-          });
-        },
-      },
-    ];
-    showContextMenu(event.clientX, event.clientY, options);
-  };
-
-  const handleThumbnailContextMenu = (event: any, path: string) => {
-    event.preventDefault();
-    event.stopPropagation();
-
-    const isTargetInSelection = multiSelectedPaths.includes(path);
-    let finalSelection;
-
-    if (!isTargetInSelection) {
-      finalSelection = [path];
-      setMultiSelectedPaths([path]);
-      if (!selectedImage) {
-        setLibraryActivePath(path);
-      }
-    } else {
-      finalSelection = multiSelectedPaths;
-    }
-
-    const commonTags = getCommonTags(finalSelection);
-
-    const selectionCount = finalSelection.length;
-    const isSingleSelection = selectionCount === 1;
-    const isEditingThisImage = selectedImage?.path === path;
-    const deleteLabel = isSingleSelection ? 'Delete Image' : `Delete ${selectionCount} Images`;
-    const exportLabel = isSingleSelection ? 'Export Image' : `Export ${selectionCount} Images`;
-
-    const selectionHasVirtualCopies =
-      isSingleSelection &&
-      !finalSelection[0].includes('?vc=') &&
-      imageList.some((image) => image.path.startsWith(`${finalSelection[0]}?vc=`));
-
-    const hasAssociatedFiles = finalSelection.some((selectedPath) => {
-      const lastDotIndex = selectedPath.lastIndexOf('.');
-      if (lastDotIndex === -1) return false;
-      const basePath = selectedPath.substring(0, lastDotIndex);
-      return imageList.some((image) => image.path.startsWith(basePath + '.') && image.path !== selectedPath);
-    });
-
-    let deleteSubmenu;
-    if (selectionHasVirtualCopies) {
-      deleteSubmenu = [
-        { label: 'Cancel', icon: X, onClick: () => {} },
-        {
-          label: 'Confirm Delete + Virtual Copies',
-          icon: Check,
-          isDestructive: true,
-          onClick: () => executeDelete(finalSelection, { includeAssociated: false }),
-        },
-      ];
-    } else if (hasAssociatedFiles) {
-      deleteSubmenu = [
-        { label: 'Cancel', icon: X, onClick: () => {} },
-        {
-          label: 'Delete Selected Only',
-          icon: Check,
-          isDestructive: true,
-          onClick: () => executeDelete(finalSelection, { includeAssociated: false }),
-        },
-        {
-          label: 'Delete + Associated',
-          icon: Check,
-          isDestructive: true,
-          onClick: () => executeDelete(finalSelection, { includeAssociated: true }),
-        },
-      ];
-    } else {
-      deleteSubmenu = [
-        { label: 'Cancel', icon: X, onClick: () => {} },
-        {
-          label: 'Confirm',
-          icon: Check,
-          isDestructive: true,
-          onClick: () => executeDelete(finalSelection, { includeAssociated: false }),
-        },
-      ];
-    }
-
-    const deleteOption = {
-      label: deleteLabel,
-      icon: Trash2,
-      isDestructive: true,
-      submenu: deleteSubmenu,
-    };
-
-    const pasteLabel = isSingleSelection ? 'Paste Adjustments' : `Paste Adjustments to ${selectionCount} Images`;
-    const resetLabel = isSingleSelection ? 'Reset Adjustments' : `Reset Adjustments on ${selectionCount} Images`;
-    const copyLabel = isSingleSelection ? 'Copy Image' : `Copy ${selectionCount} Images`;
-    const autoAdjustLabel = isSingleSelection ? 'Auto Adjust Image' : `Auto Adjust Images`;
-    const renameLabel = isSingleSelection ? 'Rename Image' : `Rename ${selectionCount} Images`;
-    const cullLabel = isSingleSelection ? 'Cull Image' : `Cull Images`;
-    const collageLabel = isSingleSelection ? 'Frame Image' : 'Create Collage';
-    const stitchLabel = `Stitch Panorama`;
-    const mergeLabel = `Merge to HDR`;
-
-    const handleCreateVirtualCopy = async (sourcePath: string) => {
-      try {
-        await invoke(Invokes.CreateVirtualCopy, { sourceVirtualPath: sourcePath });
-        await refreshImageList();
-      } catch (err) {
-        console.error('Failed to create virtual copy:', err);
-        setError(`Failed to create virtual copy: ${err}`);
-      }
-    };
-
-    const handleApplyAutoAdjustmentsToSelection = () => {
-      if (finalSelection.length === 0) return;
-
-      invoke(Invokes.ApplyAutoAdjustmentsToPaths, { paths: finalSelection })
-        .then(async () => {
-          if (selectedImage && finalSelection.includes(selectedImage.path)) {
-            const metadata: Metadata = await invoke(Invokes.LoadMetadata, {
-              path: selectedImage.path,
-            });
-            if (metadata.adjustments && !metadata.adjustments.is_null) {
-              const normalized = normalizeLoadedAdjustments(metadata.adjustments);
-              setLiveAdjustments(normalized);
-              resetAdjustmentsHistory(normalized);
-            }
-          }
-          if (libraryActivePath && finalSelection.includes(libraryActivePath)) {
-            const metadata: Metadata = await invoke(Invokes.LoadMetadata, {
-              path: libraryActivePath,
-            });
-            if (metadata.adjustments && !metadata.adjustments.is_null) {
-              const normalized = normalizeLoadedAdjustments(metadata.adjustments);
-              setLibraryActiveAdjustments(normalized);
-            }
-          }
-        })
-        .catch((err) => {
-          console.error('Failed to apply auto adjustments to paths:', err);
-          setError(`Failed to apply auto adjustments: ${err}`);
-        });
-    };
-
-    const onExportClick = () => {
-      if (selectedImage) {
-        if (selectedImage.path !== path) {
-          handleImageSelect(path);
-        }
-        setRenderedRightPanel(Panel.Export);
-        setActiveRightPanel(Panel.Export);
-      } else {
-        setIsLibraryExportPanelVisible(true);
-      }
-    };
-
-    const options = [
-      ...(!isEditingThisImage
-        ? [
-            {
-              disabled: !isSingleSelection,
-              icon: Edit,
-              label: 'Edit Image',
-              onClick: () => handleImageSelect(finalSelection[0]),
-            },
-            {
-              icon: Save,
-              label: exportLabel,
-              onClick: onExportClick,
-            },
-            { type: OPTION_SEPARATOR },
-          ]
-        : [
-            {
-              icon: Save,
-              label: exportLabel,
-              onClick: onExportClick,
-            },
-            { type: OPTION_SEPARATOR },
-          ]),
-      {
-        disabled: !isSingleSelection,
-        icon: Copy,
-        label: 'Copy Adjustments',
-        onClick: async () => {
-          try {
-            const metadata: any = await invoke(Invokes.LoadMetadata, { path: finalSelection[0] });
-            const sourceAdjustments =
-              metadata.adjustments && !metadata.adjustments.is_null
-                ? { ...INITIAL_ADJUSTMENTS, ...metadata.adjustments }
-                : INITIAL_ADJUSTMENTS;
-            const adjustmentsToCopy: any = {};
-            for (const key of COPYABLE_ADJUSTMENT_KEYS) {
-              if (Object.prototype.hasOwnProperty.call(sourceAdjustments, key)) {
-                adjustmentsToCopy[key] = sourceAdjustments[key];
-              }
-            }
-            setCopiedAdjustments(adjustmentsToCopy);
-            setIsCopied(true);
-          } catch (err) {
-            console.error('Failed to load metadata for copy:', err);
-            setError(`Failed to copy adjustments: ${err}`);
-          }
-        },
-      },
-      {
-        disabled: copiedAdjustments === null,
-        icon: ClipboardPaste,
-        label: pasteLabel,
-        onClick: handlePasteAdjustments,
-      },
-      {
-        label: 'Productivity',
-        icon: Gauge,
-        submenu: [
-          {
-            label: autoAdjustLabel,
-            icon: Aperture,
-            onClick: handleApplyAutoAdjustmentsToSelection,
-          },
-          {
-            disabled: !isSingleSelection,
-            icon: CopyPlus,
-            label: 'Create Virtual Copy',
-            onClick: () => handleCreateVirtualCopy(finalSelection[0]),
-          },
-          {
-            label: 'Denoise',
-            icon: Grip,
-            disabled: !isSingleSelection,
-            onClick: () => {
-              setDenoiseModalState({
-                isOpen: true,
-                isProcessing: false,
-                previewBase64: null,
-                error: null,
-                targetPath: finalSelection[0],
-                progressMessage: null,
-              });
-            },
-          },
-          {
-            label: 'Convert Negative',
-            icon: Film,
-            disabled: !isSingleSelection,
-            onClick: () => {
-              setNegativeModalState({
-                isOpen: true,
-                targetPath: finalSelection[0],
-              });
-            },
-          },
-          {
-            disabled: selectionCount < 2 || selectionCount > 30,
-            icon: SquaresUnite,
-            label: stitchLabel,
-            onClick: () => {
-              setPanoramaModalState({
-                error: null,
-                finalImageBase64: null,
-                isOpen: true,
-                progressMessage: 'Starting panorama process...',
-                stitchingSourcePaths: finalSelection,
-              });
-              invoke(Invokes.StitchPanorama, { paths: finalSelection }).catch((err) => {
-                setPanoramaModalState((prev: PanoramaModalState) => ({
-                  ...prev,
-                  error: String(err),
-                  isOpen: true,
-                  progressMessage: 'Failed to start.',
-                }));
-              });
-            },
-          },
-          {
-            disabled: selectionCount < 2 || selectionCount > 9,
-            icon: Images,
-            label: mergeLabel,
-            onClick: () => {
-              setHdrModalState({
-                error: null,
-                finalImageBase64: null,
-                isOpen: true,
-                progressMessage: 'Starting hdr process...',
-                stitchingSourcePaths: finalSelection,
-              });
-              invoke(Invokes.MergeHdr, { paths: finalSelection }).catch((err) => {
-                setHdrModalState((prev: HdrModalState) => ({
-                  ...prev,
-                  error: String(err),
-                  isOpen: true,
-                  progressMessage: 'Failed to start.',
-                }));
-              });
-            },
-          },
-          {
-            icon: LayoutTemplate,
-            label: collageLabel,
-            onClick: () => {
-              const imagesForCollage = imageList.filter((img) => finalSelection.includes(img.path));
-              setCollageModalState({
-                isOpen: true,
-                sourceImages: imagesForCollage,
-              });
-            },
-            disabled: selectionCount === 0 || selectionCount > 9,
-          },
-          {
-            label: cullLabel,
-            icon: Users,
-            onClick: () =>
-              setCullingModalState({
-                isOpen: true,
-                progress: null,
-                suggestions: null,
-                error: null,
-                pathsToCull: finalSelection,
-              }),
-            disabled: selectionCount < 2,
-          },
-        ],
-      },
-      { type: OPTION_SEPARATOR },
-      {
-        label: copyLabel,
-        icon: Copy,
-        onClick: () => {
-          setCopiedFilePaths(finalSelection);
-          setIsCopied(true);
-        },
-      },
-      {
-        disabled: !isSingleSelection,
-        icon: CopyPlus,
-        label: 'Duplicate Image',
-        onClick: async () => {
-          try {
-            await invoke(Invokes.DuplicateFile, { path: finalSelection[0] });
-            await refreshImageList();
-          } catch (err) {
-            console.error('Failed to duplicate file:', err);
-            setError(`Failed to duplicate file: ${err}`);
-          }
-        },
-      },
-      { icon: FileEdit, label: renameLabel, onClick: () => handleRenameFiles(finalSelection) },
-      { type: OPTION_SEPARATOR },
-      {
-        icon: Star,
-        label: 'Rating',
-        submenu: [0, 1, 2, 3, 4, 5].map((rating: number) => ({
-          label: rating === 0 ? 'No Rating' : `${rating} Star${rating !== 1 ? 's' : ''}`,
-          onClick: () => handleRate(rating, finalSelection),
-        })),
-      },
-      {
-        label: 'Color Label',
-        icon: Palette,
-        submenu: [
-          { label: 'No Label', onClick: () => handleSetColorLabel(null, finalSelection) },
-          ...COLOR_LABELS.map((label: Color) => ({
-            label: label.name.charAt(0).toUpperCase() + label.name.slice(1),
-            color: label.color,
-            onClick: () => handleSetColorLabel(label.name, finalSelection),
-          })),
-        ],
-      },
-      {
-        label: 'Tagging',
-        icon: Tag,
-        submenu: [
-          {
-            customComponent: TaggingSubMenu,
-            customProps: {
-              paths: finalSelection,
-              initialTags: commonTags,
-              onTagsChanged: handleTagsChanged,
-              appSettings,
-            },
-          },
-        ],
-      },
-      { type: OPTION_SEPARATOR },
-      {
-        disabled: !isSingleSelection,
-        icon: Folder,
-        label: 'Show in File Explorer',
-        onClick: () => {
-          invoke(Invokes.ShowInFinder, { path: finalSelection[0] }).catch((err) =>
-            setError(`Could not show file in explorer: ${err}`),
-          );
-        },
-      },
-      { label: resetLabel, icon: RotateCcw, onClick: () => handleResetAdjustments(finalSelection) },
-      deleteOption,
-    ];
-    showContextMenu(event.clientX, event.clientY, options);
-  };
-
-  const handleCreateFolder = async (folderName: string) => {
-    if (folderName && folderName.trim() !== '' && folderActionTarget) {
-      try {
-        await invoke(Invokes.CreateFolder, { path: `${folderActionTarget}/${folderName.trim()}` });
-        refreshAllFolderTrees();
-      } catch (err) {
-        setError(`Failed to create folder: ${err}`);
-      }
-    }
-  };
-
-  const handleRenameFolder = async (newName: string) => {
-    if (newName && newName.trim() !== '' && folderActionTarget) {
-      try {
-        const oldPath = folderActionTarget;
-        const trimmedNewName = newName.trim();
-
-        await invoke(Invokes.RenameFolder, { path: oldPath, newName: trimmedNewName });
-
-        const parentDir = getParentDir(oldPath);
-        const separator = oldPath.includes('/') ? '/' : '\\';
-        const newPath = parentDir ? `${parentDir}${separator}${trimmedNewName}` : trimmedNewName;
-
-        const newAppSettings = { ...appSettings } as AppSettings;
-        let settingsChanged = false;
-
-        if (rootPath === oldPath) {
-          setRootPath(newPath);
-          newAppSettings.lastRootPath = newPath;
-          settingsChanged = true;
-        }
-        if (currentFolderPath?.startsWith(oldPath)) {
-          const newCurrentPath = currentFolderPath.replace(oldPath, newPath);
-          setCurrentFolderPath(newCurrentPath);
-        }
-
-        const currentPins = appSettings?.pinnedFolders || [];
-        if (currentPins.includes(oldPath)) {
-          const newPins = currentPins.map((p) => (p === oldPath ? newPath : p)).sort((a, b) => a.localeCompare(b));
-          newAppSettings.pinnedFolders = newPins;
-          settingsChanged = true;
-        }
-
-        if (settingsChanged) {
-          handleSettingsChange(newAppSettings);
-        }
-
-        await refreshAllFolderTrees();
-      } catch (err) {
-        setError(`Failed to rename folder: ${err}`);
-      }
-    }
-  };
-
-  const handleFolderTreeContextMenu = (event: any, path: string, isCurrentlyPinned?: boolean) => {
-    event.preventDefault();
-    event.stopPropagation();
-    const targetPath = path || rootPath;
-    if (!targetPath) {
-      return;
-    }
-    const isRoot = targetPath === rootPath;
-    const numCopied = copiedFilePaths.length;
-    const copyPastedLabel = numCopied === 1 ? 'Copy image here' : `Copy ${numCopied} images here`;
-    const movePastedLabel = numCopied === 1 ? 'Move image here' : `Move ${numCopied} images here`;
-
-    const pinOption = isCurrentlyPinned
-      ? {
-          icon: PinOff,
-          label: 'Unpin Folder',
-          onClick: () => handleTogglePinFolder(targetPath),
-        }
-      : {
-          icon: Pin,
-          label: 'Pin Folder',
-          onClick: () => handleTogglePinFolder(targetPath),
-        };
-
-    const options = [
-      pinOption,
-      { type: OPTION_SEPARATOR },
-      {
-        icon: FolderPlus,
-        label: 'New Folder',
-        onClick: () => {
-          setFolderActionTarget(targetPath);
-          setIsCreateFolderModalOpen(true);
-        },
-      },
-      {
-        disabled: isRoot,
-        icon: FileEdit,
-        label: 'Rename Folder',
-        onClick: () => {
-          setFolderActionTarget(targetPath);
-          setIsRenameFolderModalOpen(true);
-        },
-      },
-      { type: OPTION_SEPARATOR },
-      {
-        disabled: copiedFilePaths.length === 0,
-        icon: ClipboardPaste,
-        label: 'Paste',
-        submenu: [
-          {
-            label: copyPastedLabel,
-            onClick: async () => {
-              try {
-                await invoke(Invokes.CopyFiles, { sourcePaths: copiedFilePaths, destinationFolder: targetPath });
-                if (targetPath === currentFolderPath) handleLibraryRefresh();
-              } catch (err) {
-                setError(`Failed to copy files: ${err}`);
-              }
-            },
-          },
-          {
-            label: movePastedLabel,
-            onClick: async () => {
-              try {
-                await invoke(Invokes.MoveFiles, { sourcePaths: copiedFilePaths, destinationFolder: targetPath });
-                setCopiedFilePaths([]);
-                setMultiSelectedPaths([]);
-                refreshAllFolderTrees();
-                handleLibraryRefresh();
-              } catch (err) {
-                setError(`Failed to move files: ${err}`);
-              }
-            },
-          },
-        ],
-      },
-      { icon: FolderInput, label: 'Import Images', onClick: () => handleImportClick(targetPath) },
-      { type: OPTION_SEPARATOR },
-      {
-        icon: Folder,
-        label: 'Show in File Explorer',
-        onClick: () =>
-          invoke(Invokes.ShowInFinder, { path: targetPath }).catch((err) => setError(`Could not show folder: ${err}`)),
-      },
-      ...(path
-        ? [
-            {
-              disabled: isRoot,
-              icon: Trash2,
-              isDestructive: true,
-              label: 'Delete Folder',
-              submenu: [
-                { label: 'Cancel', icon: X, onClick: () => {} },
-                {
-                  label: 'Confirm',
-                  icon: Check,
-                  isDestructive: true,
-                  onClick: async () => {
-                    try {
-                      await invoke(Invokes.DeleteFolder, { path: targetPath });
-                      if (currentFolderPath?.startsWith(targetPath)) await handleSelectSubfolder(rootPath);
-                      refreshAllFolderTrees();
-                    } catch (err) {
-                      setError(`Failed to delete folder: ${err}`);
-                    }
-                  },
-                },
-              ],
-            },
-          ]
-        : []),
-    ];
-    showContextMenu(event.clientX, event.clientY, options);
-  };
-
-  const handleMainLibraryContextMenu = (event: any) => {
-    event.preventDefault();
-    event.stopPropagation();
-    const numCopied = copiedFilePaths.length;
-    const copyPastedLabel = numCopied === 1 ? 'Copy image here' : `Copy ${numCopied} images here`;
-    const movePastedLabel = numCopied === 1 ? 'Move image here' : `Move ${numCopied} images here`;
-
-    const options = [
-      {
-        label: 'Paste',
-        icon: ClipboardPaste,
-        disabled: copiedFilePaths.length === 0,
-        submenu: [
-          {
-            label: copyPastedLabel,
-            onClick: async () => {
-              try {
-                await invoke(Invokes.CopyFiles, { sourcePaths: copiedFilePaths, destinationFolder: currentFolderPath });
-                handleLibraryRefresh();
-              } catch (err) {
-                setError(`Failed to copy files: ${err}`);
-              }
-            },
-          },
-          {
-            label: movePastedLabel,
-            onClick: async () => {
-              try {
-                await invoke(Invokes.MoveFiles, { sourcePaths: copiedFilePaths, destinationFolder: currentFolderPath });
-                setCopiedFilePaths([]);
-                setMultiSelectedPaths([]);
-                refreshAllFolderTrees();
-                handleLibraryRefresh();
-              } catch (err) {
-                setError(`Failed to move files: ${err}`);
-              }
-            },
-          },
-        ],
-      },
-      {
-        icon: FolderInput,
-        label: 'Import Images',
-        onClick: () => handleImportClick(currentFolderPath as string),
-        disabled: !currentFolderPath,
-      },
-    ];
-    showContextMenu(event.clientX, event.clientY, options);
-  };
 
   const memoizedFolderTree = useMemo(
     () =>
